prevent git from being called during configure from a tarball
The logic to check if someone updated the package dependency files makes a call
to git, which fails when configuring from a tarball since the tarball is not a
git repo. Making this check development mode only so that we do not have odd git
errors when configuring from the tarball.