MueLu: bugfix in Level::Release system
This fixes the temporary fix from commit
9923b2b5c778cbce39b67fff7643e5ba3ea490c6
(see email correspondence, too)

introduced a SmootherPrototypeBase class to be able
to release the dependencies of a SmootherPrototype
without making the internal RequestMode flags public.

Level::Release only releases dependencies if
the Build function of the generating factory has
not been called (i.e. Get has not been called for
all variable, (fac, ename) depends on).

To make sure, that the Request/Release system is working
properly it's highly important that all data, that is
requested in DeclareInput of a factory is also used in
the Build function (Get call).