Committing GPL script cloc.pl and some specialized lanaguage def files.
The GPL Perl script cloc.pl is such a useful tool, that I think it is
justified to add it to the TriBITS sources.  The total size is only about 350K
and therefore does not impact the size of the git repo much at all.

I also added some specialized language definition files for core compiled
languages and for script languages.  This is a nice way to look at the source
code in a TriBITS project.