Creating Teuchos::FilteredIterator to support ordered ParameterList
The class Teuchos::FilteredIterator was written from scratch to allow
filtering arbitrary bidirectional iterators based on some predicate.  This is
pretty well unit tested.

This class will be used to help support changing the Teuchos::ParameterList
class so that the parameters/sublists maintain order.

This is part of CASL VRI Kanban story #2504.

The steps to create this are given below (individual commits on a local branch):

1) Creating partial Filtered iterator implementation (SQUASH AGAINST THIS).

This implementation is fairly complete without filtering.

2) Finished FilteredIterator for usage (SQUASH THIS).

3) Changed unit test code to avoid having to print iterators.

In a non-debug version of the code, I just could not get std::vector<T>
iterators to print.  I tried everything I could think of and just could not
get it to work.  Therefore, I switched to using the testing macros specialized
for iterators (which I should have used all along).  It is a real problem that
the C++ standard does not require ostream operators for iterators.  This makes
portable debugging very hard.

4) Fixing documentation for operator<<() (SQUASH THIS).