Making ParmeterList class ordered
Here, I have replaced the guts of the implemenation of Teuchos::ParameterList
to make it preserve the order in which the parameters and sublists are added.
I turns out that the behavior of the PL class that people have come to rely on
had a pretty strong dependence on the behavior of the std::map implementation.
Through a lot of effort, I have tried to maintain all such behavior, except
for the ordering of the parameters and sublists returned by the iterators.

There are some refactorings that someone might want to do after this including:

1) Print the params and sublists in the order entered (instead of params first
then sublists)

2) Fix the behavior of getPtr() to throw on incorrect types and not just
return null.

Below is the list of local commits that I made when doing this in my local
branch:

1) Starting to replace guts of PL.

2) Adding some more unit tests.

3) More unit tests, more updates.

4) Basic iterators are working.

5) Removing old implementation some.

6) Fixing copy constructor and assignment operators.

7) Getting 100% rid of old std::map implementation.

8) Implemented getPtr() and getEntryRCP(), cleaned up a little.

9) Unit tested and fixed isType(), isParameter(), and isSublist().

10) Adding unit tests for multiple sublists to catch object no movement requirement.

11) Making getPtr() functions return null for existing param, wrong type.

Whoever write the initial version of the getPtr() functions and wrote the
tests for them made it so that requesting an existing parameter of the wrong
type returns null instead of throwing.  I am not sure if this behavior makes
sense logically (what is the use-case for this?) but to maintain backward
compatibility, I have added the unit tests and changed the behavior to
preserve it.

12) Cleaning up some formatting and removing non-longer used find() functions.

13) Fixing behavior w.r.t. validators to match backward-compatibility, etc.