Fix issue with setting the required cmake version
bug: 5536

Jonathan Hu noted that cmake 2.4.8 would still try to configure Trilinos, but
then throw an error about the syntax of an IF statement.

The issue was that the call to CMAKE_MINIMUM_REQUIRED happened after significant
amounts of code that weren't compatible with older versions of cmake. I moved it
to before the project call in the base CMakeLists.txt file which is before most
of the cmake code in tribits. However the file where tribits defines the
required version(TribitsConstants.cmake) had pieces of cmake code that aren't
compatible with very old versions of cmake(2.4.8), specifically the CMAKE_POLICY
calls. I moved the policy calls to the TribitsProject.cmake file as the first
thing done so that all tribits code will still adhere to the policies properly.
That isn't an ideal place for them since project isn't really about base tribits
setup, however, it isn't safe to call them before the call to
CMAKE_MINIMUM_REQUIRED and I didn't think it was wise to make another file that
had to be included indepently by a project to get the policy requirements. Any
project using tribits will have to include that file very early so the policies
will be set by tribits before significant amounts of tribits work is done.

There was one other minor change that had to be made to TribitsConstants.cmake.
There was a simple if that defined the git command based on whether the current
machine was windows or not. Cmake 2.4 does not support leaving out the
conditional from the ELSE and ENDIF statements that is used throughout Tribits,
including here. Since this was a legitimate constant I thought it was best to
just make that simple IF compliant to older versions of cmake instead of
breaking it out and putting it somewhere after the CMAKE_MINIMUM_REQUIRED call.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=92,notpassed=0 (5.28 min)
1) SERIAL_RELEASE => passed: passed=92,notpassed=0 (5.25 min)