Adding warning that Fortran is disabled by default on Windows builds.
I added this warning because future projects wth Fortran that use TriBITS and
want to support Windows will have to deal with this issue.  SCALE was not
configuring correctly on Windows because Fortran was being silently disabled.
Hopefully this will avoid some confusion in the future.

Note that the warning is only emitted if ${PROJECT_NAME}_ENABLE_Fortran is
being set to OFF by default on Windows (i.e. WIN32=ON).

I only tested this behavior manually and it is not unit tested.  However, this
behavior just an if statement and a message() commnand and does not affect the
behavior of the software at all, only a warning message that is printed or
not.  It is possible to unit test the setup of these options but we would have
to add some infrastructure to support this.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=92,notpassed=0 (1.31 min)
1) SERIAL_RELEASE => passed: passed=92,notpassed=0 (0.93 min)
2) MPI_DEBUG_SS => passed: passed=92,notpassed=0 (1.35 min)
3) SERIAL_RELEASE_SS => passed: passed=92,notpassed=0 (0.95 min)