Setting Optika_DO_GUI_UNIT_TESTS=OFF for all builds on the CASL fissile machines (VRI Kanban #2529)
This should allow QT to build and run the non-GUI unit tests just fine with
the modify FindTPLQT.cmake module that is now in Trilinos/cmake/tibits/tpls/.

For more details, see VRI Kanban Trac ticket #2529.

Note that Optika is still disable by default so this does not really impact
the basic VERA builds at all.  This is just done to test the FindTPLQT.cmake
module and to prepair for the possible future usage of Optika in VERA.