Adding test showing a problem in the package directory for case given in extraRepoOnePackage
This new test case is for an extra repo that has just one package at the base
level.  This turns out to be a common use case in the CASL VERA project that
uses TriBITS.  This test shows that with the current implementation the
package directory turns out to be:

    extraRepoOnePackage/.

instead of:

    extraRepoOnePackage

This seems minore but it turns out to be an important issue because of the way
that the checkin-test.py script forms directory paths before looking for
matches.

This is a failing test so you can expect the new test to fail if you build
directly against this commit.  However, the next commit will fix this test.