Handling the case where an extra repo has just one package at the base level
This fixes the case for where an extra repo that has just one package at the
base level.  This turns out to be a common use case in the CASL VERA project
that uses TriBITS (e.g. CASLRAVE, CASLBOA, DeDCART, etc.).  This fixes the
unit test I just commited where with the current implementation, the package
directory turns out to be:

    extraRepoOnePackage/.

instead of:

    extraRepoOnePackage

This seems minor but it turns out to be an important issue because of the way
that the checkin-test.py script forms directory paths before looking for
matches.

I have manually tested that this change now detects changes in such a repo
with the checkin-test.py script (a change in CASLRAVE is picked up by the
checkin-test.py script).

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=93,notpassed=0 (1.53 min)
1) SERIAL_RELEASE => passed: passed=93,notpassed=0 (1.11 min)
Other local commits for this build/test group: b46b270