fei: zero the storage for off-proc contributions instead of destroying/recreating
During matrix assemble, contributions for shared-but-not-owned nodes are
held in temporary fei storage until load-complete, then sent to owning procs
to be put into the underlying matrix. At each time-step the temporary fei storage
was being reset (destroyed) and then re-filled next time. Instead, we believe it
will be more efficient to zero the contents but leave the structure in place.
This will be tested in Stefan's scalability runs.

(From Sierra)