Update clang nightly test build directories
The clang builds were inadvertantly using the same build directories as the
default gcc builds on s903186. When I set those builds up I forgot to modify the
directory name to prevent collisions. This didn't affect the testing in anyway
as the build directories are deleted before they are used, however, it would
prevent being able to inspect the build directory for the earlier tests if
needed.