Interoperability with MPI:
Check if the host process has been pinned to a particular NUMA node.
If so then only spawn threads on that NUMA node to avoid over-subscription
of cores when MPI places one process per NUMA node.
For example when using openmpi 'mpirun --bind-to-socket --bysocket'.

Refactoring of Host internal implementation to
reduce the number of functions and depth of call stack.