Extend the mesh skinner in order to output multiple mesh files.
The CASL project would like to output both a volume mesh and
a skinned mesh.  The volume mesh is for quality control and
the skinned mesh is for coupling with other codes. In order
to do this the mesh skinner was changed slightly to include
topologies. The ReadWriteUtils were changed to take a list
of mesh parts not to output in order to skip the volume
mesh in the surface file. And the GearSkinning test was
changed to test the output of two meshes, a volume and
a surface mesh.

(From Sierra)