Disabling MueLu for the clang build
There were a number of build errors in Muelu with clang that were exposed when
the clang build was moved to nightly due to the fact nightly builds
automatically pull in the precopyright repo. Since Muelu is still in
precopyright there is no requirement on it that it has to build correctly on all
compilers and given that this is a very new compiler for us I felt it was best
to just disable Muelu for now. A bug has been filed against Muelu about this so
that we don't lose track of this issue.