Removing limitation on size of nullspace.
Previously, the nullspace size was limited to at most 63 vectors.  This was an implementation issue, not algorithmic.  The nullspace can now be any size.

This involved changing the QR fix code written by Marian Brezina.  It previously used bits and bitshifting.  I've changed it to use std::vector<bool>, which uses bits under the hood (due to a quirk in the C++ standard).  This can also be set at runtime, whereas the max size was previously hard coded.

I also took the opportunity to rename some variables with mysterious titles.