Reorganizing metric calculations.
The new parameter "compute_metrics" determines whether or not the
PartitioningProblem computes metrics after the algorithm is done.

The class PartitioningSolutionQuality computes quality metrics.  It
takes an input adapter and a solution.  For now only balance is
computed.  Graph/HG metrics will be added in the future.

Added the methods:
  PartitioningProblem::printMetrics
  PartitioningProblem::getMetrics
  PartitioningProblem::getImbalance