Correct a bug that was fixed previously, then re-appeared...
There was even a regression test to protect against this, but the
test was no good. It printed out the value in question without
checking the value for correctness.
This was reported by Tim Heister, and relates to calling globalAssemble
on Epetra_FEVector with 'Insert' instead of 'Add'. In certain scenarios
it was remembering remote values from a previous assembly cycle when
it should have reset them. It is fixed now, and so is the test.