Adding -lgmop for parts that have enabled OpenMP
This is being added to support SCALE which must have OpenMP turned on but we
don't necessarily want to turn on OpenMP for the rest of the packages since
OpenMP support is not Primary Stable code.  This will allow executables in
down-stream packages where OpenMP is not turned on to link against libraries
compiled with OpenMP turned on.

Build/Test Cases Summary
Enabled Packages: Teuchos
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=98,notpassed=0 (6.38 min)
3) SERIAL_RELEASE_SS => passed: passed=95,notpassed=0 (4.06 min)