SEACAS: Fix algebra CMakeLists.txt
This issue slipped past me earlier. Algebra suffered from a simlar issue to
exodiff where the files to build it had changed significantly in sierra, but the
Trilinos build system hadn't been updated. I used the same solution here as
well, which is to make the CMakeLists.txt more like the Sierra Jamfile and glob
all the *.f files.