Disabling TriKota when TriKota/Dakota is missing ; Disabling ForTrilinos for all projects
With this change, you can allow TriKota to be allowed to be enabled by defualt
and it will be automatically disabled when TriKota/Dakota is missig
(i.e. Dakota has not been cloned under the TriKota directory).

Also, the ForTrilinos disable was for only Trilinos.  I updated to disable for
all projects including Trilinos.