Fixed memory leaks caused by improper use of ArrayRCP::release(). It appeared the release() calls were intended to free memory associated with the array.  But in fact, release() only changes ownership of the ArrayRCP to false (whatever that means).  With ownership==false, reassignments to the ArrayRCP would not free its old memory, and that memory would be lost.
For now, I commented out the release() functions.  If we want to reduce the
maximum memory footprint of this function (as I think was the intent), we'll
have to look at rescoping parts of the function to let the ArrayRCP destructor
free the memory.

I also fixed a memory leak in the AlltoAll function; we allocated an array
and then didn't delete it.

One memory leak remains; I will track it down.