Fixing issue in Epetra adapters for model evaluator
The x_dot argument for inArgs in the Thyra epetra model evaluator was always being set to the nominal value even if the incomin ginargs had it set to null.  This is inconsistent with the stated behavior in the documentation where x_dot is an exception to the nominal values rule.  This fixes the inconsistency.