Fixing compilation for Scalar=std::complex<double>
Using Vector of type double instead of Scalar.  Its use case is simply to transfer the #nonzeros per row of the matrix.
I was using Scalar, as Epetra doesn't support size_t, and this was fine when Scalar=double.  But as this breaks for
Scalar=complex, I've now hardcode the transfer vectors to always be of type double.  This should be safe, as there
won't be any matrices with more nonzeros *per row* than can be held in a double, any time soon.