Refactor for squeaky-clean build with Tribit's aggressive '-pedantic' flag.
Fix CMakeList.txt issues.

Revision for the new View API, by example:
     View< double ***[3][3][8] , Cuda >
is a view to a rank-6 array with the first three dynamic
dimensions specified at runtime during construction.

The old API "View< double[0][0][0][3][3][8] , Cuda >"
used '[0]' to specify dynamic dimensions.  This syntax
would portably compile and provide the proper behavior.
The '-pedantic' flag caused this syntax to fail due to a
subclause in the ansi standard that states such a type
cannot declare an array.
However, this 'double[0][0][0][3][3][8]' type was only used
to compactly and concisely represent intent, and was never
use to actually declare an object.  As such the standard was
not technically violated; but the compiler assumed that a
defined type will be used as declared when enforcing '-pedantic'.

The '-pedantic' flag also causes problems for the Cuda (nvcc) compiler.
This compiler uses GNU extensions in the process of compiling C++ to ptx.
The workaround is to use the CMake CUDA option
	-D CUDA_PROPAGATE_HOST_FLAGS:BOOL=OFF
by which the Tribits-forced '-pedantic' is not passed to the Cuda compiler.
Compile options are now passed to the Cuda compiler via
	-D CUDA_NVCC_FLAGS:STRING=-arch=sm_20;-Xcompiler;-Wall,-ansi;-O3"

Fix array alignment on Host device to align on cache-line
boundaries only if the size of the to-be-aligned array sub-dimension
is larger than the cache-line.  Added cache-line query to the
Host-hwloc implementation.