Bug # 4668:  Teuchos::CommandLineProcessor::throwExceptions not working properly
This commit fixes the issue that Nico found and documented in the bug report.  It goes
one step further in actually returning a parse error through the parse() method when
throwException is not enabled.  Otherwise, the function will return a successful parsing
even though that is not true because the option was required and not found.

Since there was not a CommandLineProcessor test, I will add one.  This test has already
illustrated that there are logic issues with the CommandLineProcessor when all options
are not recognized (and ignored).  These have been fixed by this commit and the test
will ensure that the failures do not reoccur.