remove the dependence of stk_parallel on stk_env
The cyclic dependency between stk_parallel and stk_env
was the last of this kind in Trilinos.
This commit removes the dependency of stk_parallel
on stk_env. Doing so was easy enough: There was just
one call to ThrowAssertMsg() from stk_env that was
needed anyways. This is now replaced by an if(){throw}.
Also updated one dependency in connection with this.

Signed-off-by: Alan Williams <william@sandia.gov>

(From Sierra)