Framework: Make the MPI tpl disabled by default
We recently ran into an issue where a package that had a required dependency on
MPI could inadvertently enable MPI for what should be a serial build. This is
very bad behavior that is only possible because we allow MPI to be enabled or
disabled by the normal dependency logic in tribits, which is if it is unset
Tribits can decide how to set it. The solution is to make MPI "OFF" by default.
This will prevent Tribits from turning it on when doing a serial build and
shouldn't affect anyone doing an MPI build since generally you need to
explicitly set MPI to ON in the configure scripts. There is one place where this
will have an impact though, and that is the nightly testing. All of the MPI
nightly tests will need to explicitly state that they are enabling MPI a second
time so that the logic that enables packages individually will recognize that
packages that require MPI can be enabled. That change will be made in a separate
commit.