Changed Teuchos::ScalarTraits::magnitudeType to double for PCE types.
This required some specializations of a few Kokkos kernels to compute
the correct magnitude.  I also added an innerProduct() function to the
scalar traits for PCE scalar types, which allows dot products to be computed
more efficiently.   They are still promoted up to the PCE scalar type,
but otherwise they are correct and efficient.