Inserted missing Komplex_LinearProblem::UpdateValues() return statement
The Komplex_LinearProblem class has an UpdateValues() method declared
with an int return value, but the implementation of the method did not
have a return statement.  This caused a warning when compiled, but
worse, it was causing the PyTrilinos unit test suite
testKomplex_LinearProblem.py to fail on my MacBook Pro (the return
value was somehow assigned a value of 32767).

I added a return(0); to the end of the UpdateValues() method.  This
gets rid of the compilation error and the PyTrilinos unit test for
Komplex now passes.