Teuchos: Added unit tests for conversions from std::string to integer types.
I couldn't find any of the unit tests for as() or asSafe().  Grep
couldn't find them either.  Since I added new conversions from
std::string to various built-in integer types, I thought it prudent
to create my own unit tests for them.

If there are indeed unit tests for as() and asSafe(), I would welcome
consolidating my new unit tests with the existing ones.  If there are
no unit tests for as() or asSafe(), then there should be, and this is
as good a place for them as any.  Please feel free to add more unit
tests until coverage of existing code (not included in this commit)
is complete.