stk::mesh::impl  Partition can be used optionally now within toolkit.
Partition can be used for add, remove, move entities wrt Buckets,
replacing most of the code in BucketRepository.  For now, a #define
(USE_STK_MESH_IMPL_PARTITION) is required to compile the class and
use it instead of the older code.

Class passes regression tests with code inserted to be comptible with a
bug in the current Sierra code. Next step is to make sure no negative
performance impact, even though this is an interim implementation to
provide a path forward for inserting faster data structures.

(From Sierra)