Adding JFNK and user defined prec/fwd_op support to thyra group
This change allows for user defined precondiitoners and linear operators.  Previously, we pull the solver objects from the model evaluator.  A new ctor allows users to set the objects directly.  This will be used for  JFNK coupling in CASL.  The suppor tis not fully enabled (there is a block of code commented out).  This will be enabled once Ross commits changes to to the Thyra model evalutor to suppor W_prec.