Moving MSVC CLAPACK logic to BLAS and LAPACK TPL find modules (Trilinos bug 5738)
Now that Teuchos is broken into subpackages (Trilinos bug 5738) you can
configure most Teuchos subpackages (and now some of Trilinos) without needing
BLAS and LAPACK.  However, there was global logic for CLAPACK in TriBITS that
produced a warning with MS Visual C++ 2012 (MSVC).  It was annoying and
unnecssary.  In addition, the logic was such that if it found CLAPACK on your
machine, it would use it and gave the user no way to override it.  That is
bad.

Therefore, I have moved the logic for CLAPACK into FindTPLBLAS.cmake and
FindTPLLAPACK.cmake and added a conditional to allow them to be overridden.

I have tested this manually on MS Visual C++ 2012 on my Windows 7 laptop with
several different configuration cases and they seem to be working as expected.