SEACAS: aprepro - Add immutable and include functionality.
IMMUTABLE: Any variables created when the "state" is "immutable"
are constant and their value cannot be modified.  The immutable
state is set either by the comamnd line option --immutable or -X or
by putting the command {IMMUTABLE(ON)} in the input file.
When {IMMUTABLE(OFF)} is parsed, the state reverts to the default
state (immutable if aprepro started with --immutable; mutable by default).

If an attempt is made to modify an immutable varible 'a' via:
 a=b, a++, ++a, --a, a--, a+=b, a-=b, ..., then a parse error will be
detected and a message similar to:

Aprepro: (IMMUTABLE) Variable 'a' is immutable and cannot be modified (stdin, line 3)

will be output.

INCLUDE:

A file to be included prior to parsing either stdin or the normal input file
can be specified on the command line via the "-include" or "-I" option. For
example "aprepro -include pre_include_file.i normal_input_file.i"

The contents of the incldue file will be parsed and then the input file
or stdin will be parsed.  Currently, only a single include file can
be specified on the command line.

Since the -include command line option previously existed to specify an
include path to search for include files, the option value is checked and
if it specifies a directory, it is assumed to be a include path, otherwise
it is assumed to be an include file.

These capabilities were requested by the Dakota team.

(From Sierra)

Build/Test Cases Summary
Enabled Packages: SEACASAprepro
Disabled Packages: Panzer
Enabled all Forward Packages
0) MPI_DEBUG => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=3,notpassed=0 (2.92 min)
3) SERIAL_RELEASE_SS => passed: passed=5,notpassed=0 (2.62 min)