Make io-parts no longer be mechanics parts.
Region still holds a "IO_PART_HOLDER" struct, but that struct is no longer a mechanics
which means the io parts are no longer mechanics-parts. This required logic adjustments
in a lot of places in the code.

Next step will be to move the io parts from region to MeshBulkData.

(From Sierra)