MueLu: add a permutation factory class and a PermutingSmoother class
The PermutationFactory class generates a (nonsymmetric) row- and column
permutation operators P and Q for matrix A which try to maximize the
absolute values on the diagonal of A.
The permuted matrix A2 = P * A * Q^T usually is much easier to be solved
using iterative methods. This method was inspired by the work
Y. Saad, Multilevel ILU with reorderings for diagonal dominance (2004).

The PermutingSmoother class is a wrapper for the TrilinosSmoother class which
uses a permuted matrix A (generated by PermutationFactory) instead of the
original matrix. Can be used within MueLu multigrid framework.