Teuchos: Added specializations of SetScientific for built-in integers.
The specializations ensure that Tpetra::MatrixMarket::Writer and
Tpetra::MatrixMarket::Reader work for CrsMatrix specializations with
Scalar = int, long, etc.  These are perfectly legitimate Scalar types
and so it's right that we should be able to read and write matrices
of these types.

SetScientific was written to control the precision of std::ostream
output of floating-point data.  For integer data, the trivial
specialization suffices.

In the future, it would make sense for us to use the existing fields
in Teuchos::ScalarTraits to write more generic code, rather than to
supply specializations for specific integer types.  This would make
SetScientific extensible to any Scalar type for which a ScalarTraits
specialization exists, so that users could add support for a new
Scalar type just with a ScalarTraits specialization.  (For example:
Scalar = mpz_class (from GNU's GMP library), an arbitrary-size
integer type.)

Build/Test Cases Summary
Enabled Packages: TeuchosNumerics
0) MPI_DEBUG => passed: passed=13,notpassed=0 (2.02 min)
1) SERIAL_RELEASE => passed: passed=13,notpassed=0 (1.93 min)