Tpetra: Implemented private view constructor that takes Kokkos MV.
This is part of the fix for Bug 5813.  The goal is to let the
Kokkos::MultiVector do error checking and compute the view of local
data.  In order to do this, Tpetra::MultiVector needs a private
view constructor that takes a Kokkos::MultiVector (the already
constructed view of the local data), instead of the ArrayRCP and
dimensions.